# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "operationId": "ProjectsController_create",
        "summary": "Create a project",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProjectsController_list",
        "summary": "List projects (auto-scoped by role)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Planning",
                "In Progress",
                "Completed"
              ],
              "type": "string"
            }
          },
          {
            "name": "owner",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "member",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "enum": [
                "name",
                "status",
                "startDate",
                "dueDate",
                "createdAt"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "summary": "Get a single project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "summary": "Update project fields",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "summary": "Soft-delete a project (cascades to tasks/comments)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "projects"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/permission-schemes": {
      "post": {
        "operationId": "PermissionSchemesController_create",
        "summary": "Create a reusable permission scheme for this organization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionSchemeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "permission-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PermissionSchemesController_list",
        "summary": "List this organization's permission schemes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "permission-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/permission-schemes/{id}": {
      "patch": {
        "operationId": "PermissionSchemesController_update",
        "summary": "Update a permission scheme's name and/or grants",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionSchemeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "permission-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PermissionSchemesController_remove",
        "summary": "Delete a permission scheme (must not be assigned to any project)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "permission-schemes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/project-roles": {
      "post": {
        "operationId": "ProjectRolesController_create",
        "summary": "Create a reusable project role for this organization (Module 6)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "project-roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProjectRolesController_list",
        "summary": "List this organization's project roles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "project-roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/project-roles/{id}": {
      "patch": {
        "operationId": "ProjectRolesController_update",
        "summary": "Update a project role's name and/or description",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "project-roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProjectRolesController_remove",
        "summary": "Delete a project role (must not be assigned on any project)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "project-roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Project & Task Management API",
    "description": "REST API for a Jira/Trello-style project and task management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Website Redesign"
          },
          "description": {
            "type": "string",
            "example": "Revamp the marketing site."
          },
          "startDate": {
            "type": "string",
            "example": "2026-01-01",
            "description": "Defaults to today if omitted."
          },
          "dueDate": {
            "type": "string",
            "example": "2026-12-31"
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "owner": {
            "type": "string",
            "description": "Admin only: assign a different Manager/Admin as owner"
          },
          "key": {
            "type": "string",
            "example": "SUP",
            "description": "Issue-key prefix (e.g. \"SUP\" for SUP-101). Auto-derived from the name on first use if omitted."
          },
          "boardType": {
            "type": "string",
            "enum": [
              "Kanban",
              "Scrum"
            ],
            "description": "Kanban hides the Backlog/Sprint-board/Calendar tabs client-side. Defaults to Scrum."
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Website Redesign"
          },
          "description": {
            "type": "string",
            "example": "Revamp the marketing site."
          },
          "startDate": {
            "type": "string",
            "example": "2026-01-01",
            "description": "Defaults to today if omitted."
          },
          "dueDate": {
            "type": "string",
            "example": "2026-12-31"
          },
          "key": {
            "type": "string",
            "example": "SUP",
            "description": "Issue-key prefix (e.g. \"SUP\" for SUP-101). Auto-derived from the name on first use if omitted."
          },
          "boardType": {
            "type": "string",
            "enum": [
              "Kanban",
              "Scrum"
            ],
            "description": "Kanban hides the Backlog/Sprint-board/Calendar tabs client-side. Defaults to Scrum."
          }
        }
      },
...
      "PermissionGrantDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "CreateIssue",
              "Assign",
              "Transition",
              "Delete",
              "EditCustomFields",
              "ManageSprint"
            ]
          },
          "allowedRoles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "PlatformAdmin",
                "Admin",
                "Manager",
                "Developer"
              ]
            }
          },
          "allowedUserIds": {
            "description": "User ids individually granted this action",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedTeamIds": {
            "description": "Team ids granted this action (Module 6)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedProjectRoleIds": {
            "description": "Project Role ids granted this action (Module 6)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "action",
          "allowedRoles",
          "allowedUserIds"
        ]
      },
      "CreatePermissionSchemeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Strict (QA cannot delete)"
          },
          "grants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionGrantDto"
            }
          }
        },
        "required": [
          "name",
          "grants"
        ]
      },
      "UpdatePermissionSchemeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Strict (QA cannot delete)"
          },
          "grants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionGrantDto"
            }
          }
        }
      },
...
      "CreateProjectRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "QA Lead"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateProjectRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "QA Lead"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    }
  }
}